benchmarks/requirements.txt
networkx>=3.0
scikit-learn>=1.3
umap-learn>=0.5
numpy>=1.24
scipy>=1.10
